Move FunctionCompiler to a separate file
